. defines.h, iftasks.cpp, rpcif.cpp: added ERROR_USER_CANCEL, and eventUserCancel to let the task completion OnCompletion to be aware of the reason of the canceling in process.  It is used in QueuedReq mode, when the CIfInvokeMethodTask::OnComplete uses the cancel reason to determine if the next req should be started or not. That is, if the reason is eventCancelTask, the the next request will stay in the queue ( waiting for canceling by the taskgroup, and if either eventUserCancel or eventTimeoutCancel, the next request will proceed to run.
.iftasks.cpp, iftasks.h, irpthread.cpp, iomgr.cpp, utils.cpp, utils.h: added a
new method StartNewRecv to CIfStartRecvMsgTask, which will run on the task
thread, to off some of the load from the irpcomp thread. The idea is to evenly
distribute the load among the thread. The earlier setup of 2 task threads is
reduced to 1 task thread, and the irp threads has doubled its number to two
thread max. the number of workitem threads is also reduced to 1 thread.

. tasklets.cpp, tasklets.h: added a semaphore for the wakeup when the reentrancy
lock is hit. it performs better than usleep.

.ifhelper.h: fixed a bug in the DEFER_CALL that the l-value ObjPtr cannot
compile as the parameter of the call, by replacing make_tuple with the tuple's
constructor.

.rpcif.cpp, iftasks.cpp, proxy.h: added a parameter, last invoke task, to
InvokeNextMethod, so that it can know if it is time to start the next request,
in the QueuedReq mode. It is necessary when UserCancelRequest wants to cancel an
earlier request, which could be the current running task or someone in the
queue.

.rpcif.cpp, jsondef.h: added code to retrieve the propQueuedReq from the object
description file.