. iomgr.cpp, defines.h, utils.h, iftest.cpp: improved ClearStandAloneThreads and WaitThreadsQuit of CIoManager, to make sure all the threads have quit before CIoManager::Stop returns. If not, the CObjBase::GetActCount check cannot pass and the objects are released in a buggy order.
. ifhelper.h, actcsvr.cpp: renamed DEFER_CALL_ONESHOT to
INSTALL_COMPLETE_HANDLER for asynchronous user handler. The effort includes
passing the InvokeTask's task state to the user handler. But it is found to be
unnecessary.

. rpcif.cpp, iftasks.cpp, ifstat.h, propids.h: Added stateIoDone to the
CIfStartRecvMsgTask, so that StartNewRecv can tell if the current
CIfStartRecvMsgTask has finished the IO part and waiting for irp completion. If
it is true, StartNewRecv can proceed to issue next CIfStartRecvMsgTask without
wating for the previous CIfStartRecvMsgTask has finished. But the test shows it
is rare that CIfStartRecvMsgTask completes immediately. To do so, the
CIfParallelTask has its property methods revised to support property
propTaskState.

. irpthread.cpp, mainloop.h, utils.h: improved the load balance performance
between the threads of a thread pool by scanning the thread queue in double
directions alternatively.  Split the PopHead to PopHead and GetHead of
CTaskQueue, so that the GetLoadCount can give a more reliable load readings. it
is also done to the pending irp queue of the CIrpCompThread.

. proxy.h, rpcif.cpp: add RunTaskDirect to CIfParallelTaskGrp, so that the task
to add will be guaranteed to run on the current thread, unlike the AddAndRun,
which will put the task first in the pending set, and then be executed one by
one, which will give the other thread the opportunity to run the specified task.

. tasklets.h: removed the CReentrancyGuard for those tasks if IsMultiThreadSafe
test is true. The CIfRootTaskGroup and CIfTaskGroup and CIfParallelTask are
tasks without CReentrancyGuard. It should improve the concurrency to some
degree.