. Made the local keep-alive support work.
. Added CIfSvrConnMgr to track the client connection status. It is for the
purpose to fix a bug that, the keep-alive heartbeat was sent continously even
when the client is gone. The connection manager will record the sender address
of each CIfInvokeMethodTask, and listening on the module on/offline event to
keep update the connection status, so that, the Keep-Alive function will consult
the connection manager to know whether or not to send out a keep-alive
heartbeat. The connection manager also controls if the response of the
CIfInvokeMethodTask is able to send out or not.

. Added Keep-alive test case.