. Added testcase sftest for SendData/FetchData, or file transfer test. It also serves to test the multiple interfaces on a single object.
. ifhelper.h: Added macros DECLARE_AGGREGATED_PROXY and
DECLARE_AGGREGATED_SERVER for multi-interface support. Now we can have more
interfaces on an individual object. And also a new macro FORWARD_IF_CALL was
added to help to make the call to the proxy method on an interface-let.  For
detailed usage, please refer to test case `sftest'.

. rpcif.cpp: moved definitions of CopyFiles to filetran.cpp and moved
detail implementation of FetchData_Server and SendData_Server to
CFileTransferServer, to clarify that this two features are supported only by the
interface CFileTransferServer. Also added interface filtering in DoInvoke for
SendData/FetchData, and added interface labeling in SendFetch_Proxy.

. filetran.cpp, filetran.h, clsids.h, propids.h, ipc/Makefile, rpc/Makefile:
moved filetran.cpp and filetran.h from rpc to ipc. And refacted
CFileTransferServer an CFileTransferProxy for SendData/FetchData calls. It is
now a built-in interface dedicated for Send/Fetch purpose. It will support
streaming in the future, when the RPC module is done.

. serial.cpp: fixed a bug in CStlIntVector, which caused the deserialized vector
to pick up garbage values.

. stlcont.h: removed unnecessary mutex

. dbusdrv.cpp, tasklets.cpp: made one more check on the WaitForComplete on the
CSyncCallback, and made the return value more accurate by setting the return
code to CSyncCallback::GetError rather than WaitForComplete's return value. And
also CSyncCallback's eventTaskComp was updated to take the return value from
inpute parameter instead of zero all the time.

. tasklets.h: Put the GetParamList and GetIrpFromParams under the guard of the
task lock in the CThreadSafeTask.