. counters.cpp, counters.h, clsids.h, propids.h, ipc/Makefile, ipc/dllexp.cpp, sfdesc.json: added server/proxy implementation for interface CStatCounters. Just a prelimiary implementation, with 3 counters:     propMsgCount: invoke requests received.     propMsgRespCount: number of response messages     propMsgOutCount: number of request/event messages sent
. defines.h, proxy.h, iftasks.h, iftasks.cpp: defined and added support for the
message filters. The message filters will be invoked when the message arrives or
leaves.  The filter can return STATUS_PENDING, STATUS_SUCCESS, or
ERROR_PREMATURE. On pending, the request will resume when the filter gets back,
and on success, the next filter will take over or move on to invoke. On
ERROR_PREMATURE, the invoke or send will not happen, and the call returns with
error.

. ifhelper.h: improved the CAggregatedObject's definition on the virtual methods
by adding helper macros DEFINE_VIRT_METHOD_IMPL_NOARG and
DEFINE_VIRT_METHOD_IMPL. The two macro helps to interate the aggregated interfaces'
implemention of the virtual methods which override the version from the virtual
base class.

. iftasks.cpp: fixed a minor bug, to clear the return values of the child tasks
of a CIfTaskGroup on retry.

. ifsvr.cpp, iftest.cpp, sfsvr.h, sftest.cpp: made improvement to the inline
comments.