. clsids.h, ifhelper.h, sfdesc.json: added an built-in interface IUnknown for EnumInterfaces method. So that the user does not have to know the interface name to which the EnumInterfaces belongs. and now the major interface for CAggregatedObject has an empty handler/proxy map.
. iftasks.h, iftasks.cpp: added CIfParallelTask::OnEvent to lock the parameter
preparation and the call to operator() all the way, so that, there is no more
race condition in changing propParamList.

. defines.h, dbushook.cpp, evloop.cpp, evloop.h, gmloop.cpp, mainloop.cpp,
sevpoll.cpp, tcpport.cpp: made changes to have the watch callback tasks's
OnEvent as the main entry when the watch is triggered instead of operator(),
which has some limitations in adding stream support.

. sftest.cpp: added Pause test to this test case to check if the server can exit
gracefully