. stream.h, stream.cpp: added streaming interface. It is yet to debug, ;).
. objbase.cpp, defines.h, tasklets.cpp, tasklets.h: added Sem_Wait to handle
sem_wait in a more robust way

. clsids.h: added clsids of CIoWatchTaskProxy and CIoWatchTaskServer for stream
support.

. ipc/Makefile: added stream.cpp to the make list

. evloop.cpp: added handle of the SOURCE to the source callback's parameter list

. filetran.cpp: added interface id to the SendData/FetchData's data description
for SendFetch_Proxy to check and set interface name for the request bound to the
server.

. ifhelper.h: added helpers ::InterceptCallback and ::RemoveInterceptCallback to
reduce the code redudency.

. ifhelper.h: added FetchData implementation to macro DECLARE_AGGREGATED_SERVER,
so that, it can pick the proper override of FetchData_Server from the
interface CFileTransferServer or IStream if both are supported by the object.

. iftasks.h, iftasks.cpp: added GetIid to CIfInvokeMethodTask so that the
FetchData can tell which interface the request is from. The reason is that both
CFileTransferServer and IStream Server support FetchData request.

. rpcif.cpp, proxy.h: added support for IStream to the FetchData/SendData,
mainly to filter if the requests come from the proper interfaces,
CFileTransferServer or IStream so far.