. stmtest: Added new test case stmtest to test streaming interface. For detailed infomation, please refer to stmtest/Readme.txt
. stream.cpp, stream.h: Added CSendQue::Send to wrap the `send' which replaces the
previous `write' to prevent SIGPIPE exit.

. stream.cpp, stream.h: Add RunTask to CIoWatchTaskProxy to initialize the
m_pReqTask. Previously this value is not initialized can cause the proxy wait
endlessly even the CIfIoReqTask has already completed.

. stream.cpp, stream.h: Add abstract method OnConnected to CStreamServer to
notify the new stream comes when FetchData_Server is invoked.

. stream.cpp, stream.h: Made CStreamServer::OnStmRecv an abstract method so that
the user must implement this method in the subclass of CStreamServer.

. stream.cpp, stream.h: Added code to stop the watches in CloseChannel and
CancelRequest, since if not stopped, the watches will continuously report I/O
error.

. stream.cpp, stream.h: Added StartStream as the start point of the stream
communication instead of OpenChannel in CStreamProxy, to simplify the usage.

. stream.cpp: Refacted the CStreamProxy's OpenChannel because the earlier
implementation did not get the right pointer to CIfIoReqTask, and caused both
proxy cannot quit gracefully.

. stream.cpp: Added code to RemoveChannel in ReleaseChannel(), which caused
memory leak because the CIoWatchTask is still kept in the IStream's watch map.

. stream.cpp: fixed a bug that CIoWatchTask::operator() did not return G_SOURCE_XXX
on eventIoWatch, and caused the socket watch been wrongly removed, and
communication broken unexpectedly.

. stream.cpp: fixed a bug that the CIfParallelTask enters stateStarted rather
than stateConnected, and the eventIoWatch event was wrongly handled by base
classes because of wrong check of the task state, and cause the CIoWatchTask to
enter the stateStopped state.

. stream.cpp: fixed a bug in CSendQue::OnIoReady that the size of the data
packets is not sent out in network byte order and caused the connection broken
prematurely.

. test/Makefile: added test case stmtest

. filetran.cpp, filetran.h: Removed oboselete code.

. ifhelper.h: Added partical specialization for CAggregatedObject for
CInterfaceServer, And moved the unique FetchData_Server definition from
DECLARE_AGGREGATED_SERVER to the template class, so that the overrides of
FetchData_Server from different interfaces can exist at the same time.

. proxy.h, rpcif.cpp: Moved CustomizeRequest to the CRpcServices's RunIoTask, to
avoid more than one occurrences from  different RunIoTask callers. it is moved
right before the task is submited to run, when the request contains the complete
information for a remote call.

. rpcif.cpp: fixed bugs in the SendData/FetchData related code, including added
propCallOptions to the remote Data Description, and added propTaskId to both
local and remote Data Description for Canceling purpose and Keep-Alive.