. Added new KeepAlive support to the IStream interface, which is more reliable and lightweight.
. iftasks.cpp, iftasks.h, rpcif.cpp, stream.cpp, stream.h: Added stream specific
keep-alive support, which is the ping-pong operation through the stream channel
between proxy and server. Both sides will reset the its timer when the ping-pong
packets arrives. Thus, we have added ResetTimer methods to both
CIfInvokeMethodTask and CIfIoReqTask, and removed the CF_KEEP_ALIVE from the
stream's FETCH_DATA request, and added keep-alive timer in the
CIoWatchTaskProxy, to periodically initiates the ping-pong operation.

. ifhelper.h: fixed a bug, that CDeferredCallBase::operator() should not fail on
non-parameter calls, and another bug that DEFER_CALL cannot compile with empty
parameter list.

. rpcif.cpp: fixed a bug that CInterfaceProxy's OnKeepAliveTerm should return 0
instead of forcing STATUS_PENDING to complete the caller, CIfInvokeMethodTask.
With STATUS_PENDING, the invoke task is not released an piled a lot in the task
group.