. objbase.cpp: made change to start the objid from a random number to make sure the object id will not collide on the server side and makes the CancelRequest to cancel the wrong request.
. defines.h, rpcif.cpp, rpcpdo.cpp, prxyfdo.cpp, prxyport.h, dbusport.h: added a
flag to `module on/offline event' to notify if the event incoming is one of the
interested event the interfaces are waiting for. Now the proxy side will ignore
those events not interested, and the server side will handle all the `module
on/offline event', to update its connmgr. ScheduleModOnOfflineTask from
CRpcPdoPort and CDBusProxyFdo have one of the parameters changed from bool to
guint32.

. stream.cpp, stream.h: added StartTimer and StopTimer to CIoWatchTaskProxy, to
reduce the code redundancy. Also added ChangeState to CIoWatchTaskProxy to make
sure the StopTimer can take effect without being undone by AddTimer from the
CIoWatchTask::OnIoReady

. stmsvr.h, stmtest.cpp: added synchronization for WriteStream, when it returns
STATUS_PENDING, which indicats the data is pending in the sending buffer waiting
for transmission. The OnSendReady will notify the buffered data are sent and
ready for the new data.