. dbusprxy.cpp, fdodrv.cpp: fixed some bugs in CProxyPdoConnectTask to enable retry for 3 times when the server is not up.
. echodesc.json, driver.json: updated for RPC debugging

. dbuserr.cpp: refacted the ErrnoFromDbusErr for better performance.

. registry.cpp: fixed a bug that the registry can crash due to the RemoveDir is
not locked properly

. dbusport.cpp, dbusport.h: fixed some bugs in ip address conversion function
Ip4AddrToBytes ,BytesToString, BinToHexChar, which produce wrong port id in
proxy port.

. dbusport.cpp: replaced the implementation of CDBusBusPort::GetChildPorts with
super::GetChildPorts to obtain the right number of child ports.

. dbusport.cpp: fixed two bugs in create CDBusProxyPdo, including missing dbus
connection, and empty cfg object as NewObj's parameter.

. driver.json, rpcpdo.cpp: fixed a bug that the proxy hangs because the smaller
number of processors than the physical processors made the CPortState's state
stack to reach the upperlimit and returns -EAGAIN to DispatchData of
CRpcBasePort. Also to handle the -EAGAIN situation, the CanContinue check is now
put in a loop till the return value is not -EAGAIN.

. drvmgr.cpp: fixed a bug that port driver is not notified when StopDrivers is
called, and caused CProxyFdoDriver unable to remove the CDBusProxyFdo from the
registry.

. filetran.cpp: fixed a bug by bringing forward the obtaining of the callback ptr
than checking the return code, which is used to synchronize with the proxy. the
bug caused the proxy to wait endlessly.

. iftasks.cpp: fixed a bug that in CIfOpenPortTask, the eventPortStartFailed
cannot be handled by OnPortEvent( iEvent, hPort );

. iomgr.cpp, port.cpp: Moved pIrp->MarkPending from Port's SubmitIrp to the
IoManager's SubmitIrpInternal to fix a bug when the port stack has two ports,
and SubmitIrpInternal send the irp to the bottom port directly, the irp was
mistakenly not marked pending.

. port.cpp: fixed a bug in DetachFromPort and AttachToPort, where the lower port
ptr was not set to the right port.

. port.cpp: added property removal for the following properties,
propLowerPortPtr, propUpperPortPtr, propDrvPtr, propBusPortPtr.

. port.cpp: fixed a minor bug in SubmitStopIrp, when the return value was set
by SetPortStateWake, which is boolean value but not gint32, and the return code
turned out to be 1, an invalid error code.

. tasklets.cpp: fixed a bug that the retry count is not consistant with
GetRetry. The bug made the tasklet dangle around and cannot move forward.

. dbusprxy.cpp, prxyport.h: removed CDBusProxyPdo's CanContinue, and moved the
IsConnected test to each ctrl code handler, because it has side effect to
CRpcBasePort's DispatchData.