. rpcroute.cpp, rpcroute.h, rpcif.cpp, proxy.h: Changed the CRpcRouter to the subclass of CAggInterfaceServer to have better life cycle management.
    . Added the following methods to CRpcInterfaceBase, OnPreStop and DoStop,
    for stopping the bridges and proxies of the router gracefully, and removed
    the old Stop and StopEx accordingly. To be more specific,the StopEx
    now allows OnPreStop to complete asynchronously and resume sometime later
    with DoStop.

    . Added AddStartTasks to allow the subclass to add startup tasks at the
    interface object starting stage. One scenario is to allow the SessionManager
    to login before normal communication can start.

    . Changed the filter container from std::set to std::queue to guarantee the
    order of the filter execution fixed.

. propids.h, rpcif.cpp: added propRouterRole to differentiate the mode the
router is running in, 01 as proxy side forwarder, 02 as server side bridge, and
03 for both.

. proxy.h, rpcif.cpp, rpcroute.cpp, reqfwdr.cpp: added helper RunManagedTask for
tasks to run under the taskgroup mechanism.

. rpcroute.h: Changed the CRpcRouter to have seperate maps for Bridge Proxy
object and Bridge object. The proxies are mapped to the ip addresses, and the
bridges are mapped to ip addr plus the tcp port number. So far, we don't allow
the client to connect to the server via more than one connection.

. tcpfido.cpp, bridge.cpp: Added AllocIrpCtxExt to hold the Stream id the data
will go through. Previously all the requests will go through two default
streams, 0 and 1. and now is changed to go through the single stream 1 only
instead.  Further more, the SessionManager will send its requests via another
stream 2. That is, we now require all the requests sent down to tcpfido must
have a stream id.

. tcpfido.cpp: now the match object to register will be stored in the stream
object.  only one match for one stream object. It was discarded in the earlier
implementation.

. reqfwdr.cpp: redesigned the CReqFwdrOpenRmtPortTask's control flow.

. Tons of changes in rpc are still on the way.