. port.cpp, iomgr.cpp: added CBusPortStopSingleChildTask, CGenericBusPort::ClosePort, CPort::Unloadable and CIoManager::IsPortNoRef for dynamically unloading the port stack not referenced any more. It is a preparation for dynamically restart an interface and the port associated. The iomanager will check if the port is referenced by any interfaces when the CIoManager::ClosePort is called and will try to unload the no-used `Unloadable' port. So far the bus ports, the CDBusLocalPdo and CDBusLoopbackPdo are not Unloadable ports and cannot be unloaded.
. tcpport.cpp, frmwrk.h, iomgr.cpp: added GetCmdLineOpt and SetCmdLineOpt to
CIoManager to pass in the command line options to the internal components. And
added LoadPortOptions and propListenSock to CRpcTcpBusPort to get the
information of whether or not open the listening port to service incoming RPC
calls.

. dbusprxy.cpp, dbusport.h: added NotifyRouterOffline and OnModOnOffline to
CDBusProxyPdo to handle the router offline event. That is, to broadcast an
eventRmtSvrOffline message to the connection point on the io manager, so that
the subscribers, usually the interface objects will pick up the event and start
the stopping process.

. rpcif.cpp: Deferred the call StopEx to leave the message receiving thread,
when received the eventRmtSvrOffline. And changed the OnAdminEvent to handle
the Pause/Resume only on server side.

. iomgr.cpp, ptifmap.cpp: added GetPortPtr to CIoManager and CPortInterfaceMap,
so that the port pointer can be retrieved safely by handle. PortToHandle is
somewhat risky when we are not sure if the port exists or not.

. port.cpp: when GetProperty of the Port object along the port stack, we now
first unlock the upper port and then lock the lower port to avoid the lock
nesting.

. dbusport.cpp, port.cpp: added BuildPdoPortName to CDBusBusPort for the
CDBusProxyPdo's port-id generation. previously we used ipv4 address as the port
id. now we changed to use the guint32 as the port and a ipaddr-to-portid map for
in the CDBusBusPort for port lookup. It is due to the possibility the address
could be ipv6 or some other format, when the port-id of guint32 cannot hold.

. stmpdo.cpp, tcpport.cpp: replaced the propSrcIpAddr and propDestIpAddr
occurances with propIpAddr in CTcpStreamPdo and CRpcTcpBusPort. And made the tcp
port to listen/connect configurable. Previously the consideration of the design
is to make the single stream connection to serve the proxy/servers on both side
at the same time. It turned out to be complicated because the reqfwdr router is
per ip address, but the bridge router is per connection, as ip address and port
number.  It is not a one-to-one mapping. So it is complicated to pick from many
connections a connection to the double-direction communication. So now the
single connection is dedicated to just one direction communication, that is, one
end is fixed to be proxy and one end is fixed to be server. Note that, the
router can host both type of connections still.

. stmpdo, tcpport.cpp: Now the port id for a stream pdo is a guint32 value from
NewPdoId instead of the previous ipv4 addr, for the purpose to accomodate
addresses of both ipv4 and ipv6.

. reqfwdr.cpp, rpcroute.h: added OnCancel to CReqFwdrOpenRmtPortTask, to support
task canceling from the taskgroups.

. reqfwdr.cpp, rpcroute.h, rpcroute.cpp: added CloseRemotePort to
CRpcReqForwarder to close the bridge proxy on client request if the reference to
the bridge proxy drops to zero. Accordingly, added RemoveBridgeProxy to the
CRpcRouter to remove the registry of this bridge proxy in the router. And alse a
CReqFwdrCloseRmtPortTask is added to call the bridge proxy's Shutdown to
physically stop the interface.  the client request happens when a client
interface proxy stops and is closing the underlying proxy pdo.

. reqfwdr.cpp, rpcroute.h, rpcroute.cpp, bridge.cpp, ifstat.cpp, rpcif.cpp:
added one more parameter to the OnRmtSvrEvent, the handle of the port, so that
the method can have some more information to do the job.

. rpcroute.cpp, rpcroute.h, reqfwdr.cpp: added RemoveLocalMatchByAddr and
RemoveRemoteMatchByPortId to the CRpcRouter, which will be used in the newly added
tasks CRouterStopBridgeProxyTask and CRouterStopBridgeTask. The
RemoveLocalMatchByAddr is called when the router is about to remove the
CRpcTcpBridgeProxy when the tcp connection is lost unexpectedly.

. fidodrv.cpp: added CRpcTcpFidoDrv for creation of CRpcTcpFido during building
port stack for CTcpStreamPdo.

. iftasks.cpp: made changes to CIfParallelTaskGrp and CIfTaskGroup to enhance
the two to be multi-thread safe. That is, we are now using propNoResched as a
reentrency lock to prevent RunTask to be run on different threads at same time,
since that will damage the integrity of the information in the task group. we
are doing this as a preparation for the implementation of interface restart, and
interface removal on stop of CIoManager.

. ifhelper.h, rpcroute.h: moved _DummyClass_ and PackParams from rpcroute.h to
ifhelper.h. And replaced the VecToTuple with PackParams which greatly simplified
the CMethodProxy and CDeferredCall.