. rpcroute.cpp, rpcroute.h, tractgrp.h: Refacted the remote match addition/removal process.  unlike the earlier design, we simplified logics in CRouterAddRemoteMatchTask, CRouterEnableEventRelayTask, and CRouterStartReqFwdrProxyTask by removing the error handling code and the hooking code to start the next tasks, which are now handled by CIfTransactGroup instead. Also added 4 helper methods to facility this task, BuildStartStopReqFwdrProxy, BuildEventRelayTask, BuildAddMatchTask, BuildStartRecvTask.
. rpcroute.cpp, rpcroute.h, reqfwdr.cpp: Added OnRmtSvrEvent, OnRmtSvrOffline and
OnRmtModEvent to the CRpcRouter. The methods will generate a dbus message on the
bridge side and forward it to the interested client over the tcp connection.

. configdb.h: replaced the guint32 to intptr_t in SetIntPtr/GetIntPtr to be
workable on both 32bit and 64bit.

. iftasks.cpp, iftasks.h: added InsertTask to CIfTaskGroup, so that the task can
be added to the head of the task queue. Also added helper SetRunning to class
CIfTaskGroup to flag the taskgroup has its tasks running.

. iftasks.cpp: Replaced CanRetry to Retriable to simplify the code.

. msgmatch.h : For added destination check in IsMyReqToForward from
CRouterLocalMatch.

. rpcif.cpp: Made CRpcServices::StartEx able to be called independently by adding
state change check. And Also made change to the CRpcServices::OnEvent to make
sure the event queue exhausted before return.

. bridge.cpp: renamed CBridgeAddRemoteMatchTask to CRouterAddRemoteMatchTask and
moved the class definition to the rpcroute.cpp.

. bridge.cpp, reqfwdr.cpp: populated propPortId to the CIfIoReqTask and
CIfInvokeMethodTask, so that on shutdown of a bridge, the requests from the
bridge pending in the CRpcReqForwarderProxy can be canceled gracefully.

. reqfwdr.cpp: brought forward the AddLocalMatch and RemoveLocalMatch to
EnableDisableEvent, so that, if duplicated addition of the match will only
increase the reference count instead of a full match addition on both bridge and
reqfwdr.

. reqfwdr.cpp: added OnRmtSvrEvent, OnRmtSvrOffline and GetParallelGrp to
CRpcReqForwarderProxy to handle the bridge shutdown event. That is, to cancel
the the pending requests from the bridge.