. registry.cpp: fixed a bug in CRegistry::RemoveDir which removed both the directory and its parent directory
. iomgr.cpp, defines.h, frmwrk.h: added eventHouseClean handler to iomanager to
clean up the completed one-shot threads, and accordingly a timer `m_iHcTimer' to
CIoManager for this task.

. iomgr.cpp, irp.h, irp.cpp, pnpmgr.cpp: added irp flag IRP_CALLBACK_ONLY for
irps dedicated for synchronous purpose. That is the irp is just to coordiate
with the assoc slave irps, and it should not be completed by the port stack. it
is used in CPnpMgrStopPortAndDestroyTask.

. iomgr.cpp: made changes to unload the port if it is unloadable without
references from any interface.

. namespc.h, dbusprxy.cpp: added DBUS_DESTINATION2 to generate the destination
string by replacing the `/' with `.' from the object path.

. pnpmgr.cpp: fixed a bug which broad the remote server events to connection
point propDBusModEvent, which caused CRpcRouter unable to receive the
eventRmtSvrOffline.

. port.cpp: fixed some bugs in SubmitStartIrp and CompleteStartIrp, to align the
port state correctly in the start state-machine. and removed the wrong code to
send out eventPortStopped when the port failed to start.

. port.cpp: fixed a memory leak by removing port from the handle map when the
port stack is destroyed.

. port.cpp: fixed a bug in CompleteStopIrp to avoid sending eventPortStopped
twice.

. portdrv.cpp: fixed a typo bug, which failed to remove the port from the
driver's port map and caused memory leak.

. proxy.h, rpcif.cpp: fixed a bug by changing the RunManagedTask's 2nd parameter
type from `bool' to `const bool&' because the CBuffer is still unable to map to
bool correctly.

. rpcif.cpp: fixed some bugs in dbusproxy port creation.

. reqfwdr.cpp: fixed some bugs in CReqFwdrOpenRmtPortTask.

. stmpdo.cpp: Overrided the GetProperty and SetProperty to fail the query for
for propSrcDBusName and propDestDBusName, because the CTcpStreamPdo does not
    support the two properties.

. stmpdo.cpp, tcpfido.cpp: fixed the bug of empty bus pointer in
CTcpStreamPdo's constructor, and made changes to set the port type correctly for
CTcpStreamPdo, and CRpcTcpFido.

. stmpdo.cpp, tcpfido.cpp: to FireRmtSvrEvent at the right place, that is, only
in the OnPortReady of the topmost port of the port stack.

. stmpdo.cpp: added code in AllocIrpCtxExt to call CPort::AllocIrpCtxExt for
unknown IRP function, to fix the crash during IRP_MN_PNP_START.

. tcpport.cpp: made changes to set the watch stopped at start. and start the
read/write watch when the socket is connected successfully, to prevent wrong io
events before the socket is connected.

. tcpport.cpp: fixed a bug by detecting the connection lost when `read' returns
zero, besides POLLHUP event, which was ignored previously and caused the
mainloop in constant busy loop.

. tcpport.cpp, tcpport.h: added AddConnTimer and RemoveConnTimer to
CStmSockConnectTask for unsuccessful connect request to cleanup the resources.