rpcroute.cpp, rpcroute.h: made change to CRpcRouter::OnModEvent to filter off the unique name, which is ignored for bridge side module online/offline events.
rpcroute.cpp, reqfwdr.cpp, bridge.cpp: added ClearRemoteEvents to
CRpcTcpBridgeProxy and CRpcTcpBridge to clean up the context on client offline.
Also BuildDisEvtTaskGrp is taken out from the RunDisableEventTask for both
references from DisableRemoteEvent and ClearRemoteEvents. The CRpcReqForwarder
is the initiator of the ClearRemoteEvents on detection of the client module
offline.

tcpport.cpp: fixed a memory leak by the sock timer which was not released and
could continue to work even after the CRpcStreamSock is stopped.

pnpmgr.cpp, port.h, port.cpp: moved StopChild from PreStop to OnPrepareStop in
CGenericBusPort, when the port is still in the `READY' state and I/O are still
allowed for the child ports. Accordingly SubmitStopIrpEx, CompleteStopIrpEx and
OnPrepareStop is added to CPort for this purpose.

pnpmgr.cpp, port.cpp: made changes to reset the irp timer when the irp is made
association to the slave irps. Also the timer for a bus irp as the master irp
should be determined by the # of generations of ports.

iomgr.cpp: moved the mainloop to the last one to stop, so that to allow the
remaining I/O to complete, and the `stop' task can also complete with timeout if
failure occurs.

dbushook.cpp: added code to CDBusIoCallback's DoRemoveWatch and operator() to
avoid dbus_watch_handle get called after the watch has been removed.

pnpmgr.cpp: on eventDBusOffline, CIoManager::Stop is also called, but it could
cause endless wait because the call is on the context of a task thread, which
is expected to quit after the call, a dead-lock situation. Now we schedule the
task on a stand-alone task for the full stop.

dbushook.cpp, dbusport.cpp: Due to the mainloop is now the last one to stop, the
dbus's disconnection can be received even after the dbus port has been removed
via static function DBusMessageCallback. To fix it, the CDBusDispatchCallback's
operator added check to fail the dispatch if the CDBusIoCallback has been
removed and the CDBusBusPort is guarded by a ref count when the dbus holds a
reference to it by the parameter of DBusMessageCallback.

reqfwdr.cpp: fixed a bug that the dbus's bool type is 4 bytes wide and the c++'s
bool is one byte wide, which could make dbus_message_append_args to fail if not
converted properly.

rpcroute.cpp: for CRouterAddRemoteMatchTask, there is the need that besides the
success to tell the parent `AND' taskgroup to proceed and error to stop, to
succeed and stop, equivalent to an `OR' taskgroup. So we added a ChangeRelation
to the CIfTransactGroup to change the task's parent taskgroup from `AND' to `OR'
if the above is the case.

stmpdo.cpp: added m_bStopReady to indicate if the OnQueryStop has been called
once, and return immediately if the OnQueryStop has been called. It is because
OnQueryStop will change the port state, and duplicated OnQueryStop could result
in wrong state.

tcpport.cpp: added check in the CTcpStreamPdo's StartTimer to prevent the timer
is added when the port is already stopped. Also in CRpcListeningSock's Connect,
added `SO_REUSEADDR' to allow the address can be resued immediately after
restart.